message sends wont loop forever anymore on unknown twitter api errors, they just fail. sending delays now notify the client so an indicator can be displayed in the UI moved MessageHistory class from TwitterFollowerDB to MessagingCampaign.ts where it belongs added indexes to message history tables fixed a double-query that was happening on startup in the Client
new behavior on the client to ensure that as rows at the top of the list are removed as message are sent, that rows on the bottom can fill in. this is done by querying 1000+the number of displayed rows and storing the to-be-displayed rows in a separate 'extraRows' array.

client will now also show a notice if sending gets delayed due to api rate limits

fixed bug where show/hide contacted button was showing the wrong text